Fix php-mode issue
The php-mode stopped working recently and I've just figured out why.  In
the commit ce90255--that solved the problem with initializing features
that are not available in the system--I forgot that the php-mode has
NEVER been provided by the system, and thus it does not belong to the
modes that are supposed to be wrapped into the featurep condition.