Define hooks for git commit messages
Since git-commit-mode is a *minor* mode and text-mode happens to be
selected as major mode, it took me a while to figure out that I need to
add hooks to text-mode.

Meanwhile I noticed that max lenght for commit messages is set to 62
instead of recommended 50 characters.  I fixed the lenght to 50, so now
my settings are compliant to 50/72 rule.  Yay!