Define git-commit hook via git-commit-setup-hook
It turned out that the text-mode-hook is not the right place to define
git-commit hooks.  Many modes such as org-mode are derived from the
text-mode and they inherit this hook as well.  For instance, I don't
want to use fci-mode with org-mode.  Proper way to do this is via
git-commit-setup-hook.