Fix package-selected-packages list order
I reverted the package-selected-packages customization option and then
ran package-refresh-contents command to refresh the list to set the list
to alphabetical order.  Goal was to make this list independent from the
order in which the packages had been installed, thus the variable will
be the same on all my machines.