Fix call of CLISP function in config file
I have learned that (first ...) is a CLISP function that for some reason
works in Emacs installed at arche, but doesn't work at thor, though it
is supposed to be the same version of Emacs.  Anyway, it is a bad
practice to call CLISP function from the Emacs Lisp code, so I replaced
it with (nth 0 ...) function.