Add yaml-mode hook function for ansible minor mode
I wanted to invoke ansible minor mode as an add-on to yaml-mode, but
only for files with .yml extension.  To achieve this goal, I defined a
new hook function for yaml-mode as part of the 'ansible-extras' package.
The hook function invokes ansible minor mode according to the extension
of file (or buffer) name.