Optimize buffer killing and org tables
Default behavior of the C-x k keystroke is to ask for the name of buffer
that is going to be killed, while offering the current buffer as the
default one.  I always use this keystroke to kill current buffer, so I
have decided to optimize the process by associating the keystroke with
the kill-this-buffer function.

At the same time I have realized that it is always useful to highlight
current line when editing org tables.  I found an elegant way how to
activate the hl-line-mode automatically once cursor enters the region of
an org table.