Add adaptive-wrap and remove flycheck-grammarly
To achieve the same behavior of my LaTeX source as the one in the
Overleaf, I need to add also adaptive-wrap package.

The flycheck-grammarly package is rather slow and has limited
functionality in comparison to the actual Grammarly in the browser.  It
is more efficient to perform Grammarly check ex-post in Overleaf.