Treat adwaita theme as safe theme
For some reason when I click on Save Theme button in Customize SHA of
the selected theme file is added to the list of custom-safe-themes.  I
am not sure if it is required by any part of Emacs, but just to be on
the safe side, it is better to keep it there.