Change the key bounded to kill-this-buffer
Original intention was to replace kill-current-buffer (default function
associated with C-c k keystroke) with the kill-this-buffer function.
Later I have realized I do not want to diverge from the default
keystrokes that much.  The accelerator keystrokes in Emacs are usually
mapped to the Super key and since s-k does not have any conflicts with
the GNOME keybindings, it makes it perfect candidate for being mapped to
kill-this-buffer.  This is not only shorter, but it releases default
mapping of the C-c k back to the kill-current-buffer function.