Remove key binding for kill-this-buffer
The kill-this-buffer was designed to be executed by the File → Close
menu entry and the function is bound to menu-bar-mode.  I was about to
map s-k to kill-current-buffer, a drop in replacement of
kill-this-buffer that is not bound to menu-bar-mode.

However, according to Mickey Petersen's wonderful "Mastering Emacs"
book, Emacs power users rarely close buffers after visiting because,
unlike browser tabs, opened buffers do not really create significant
overhead.  Contrariwise, Emacs was deliberately designed *not* to have a
quick keystroke to kill current buffer.  Indeed, I found myself
accidentally killing current buffer instead of killing a sentence as
Meta key and Super key are located next to each other.