Remove projectile package
As I am getting more familiar with the integrated Emacs tooling, I found
out that I do not really need the projectile package, because current
version of project.el covers eveything I need.  This change also solves
the problem with the conflict of the s-p keybinding with GNOME default
binding that can be now released.