Disable selection compatibility mode
I was able to operate without these two selection compatibility modes
for a long time and according to Mickey Petersen, they were added only
because newcomers are expecting this behavior from a text editor.

I turned the delete-selection-mode on about a year ago, because it
looked handy and time saving.  However, in the real world it does more
harm than good.  I found myself using undo much more than before because
of accidental overwrites of the selected text.

Honestly, I never got used to the shift-select-mode, which somehow
became on by default in Emacs over the past years.  It can potentially
interfere with the other modes and I do not use this kind of selection
in the other editors.