Set default fill column to 79
Recently, I noticed that Emacs wraps text at 79 characters when the
buffer window is 80 character wide and this applies to both terminal and
graphical frame.  My original choice to wrap lines at 80 characters
started to feel a bit unnatural within the Emacs ecosystem.  As I
learned that the PEP-8 standard limits lines to 79 characters exactly
for this reason, since Guido van Rossum uses Emacs and he managed to
enforce his personal preference as a coding guideline.

I had originally an intention to enlarge widen frame size to 81
characters to accommodate default frame to my personal preference, but I
do admit that this would not work play well with the terminal.  I
realized that the most natural thing to do is to simply adopt Guido's
standard.