Add workaround for the frame resize bug
I have discovered an Emacs bug in Fedora 40.  When Emacs frame is
maximized to the left (or right) in GNOME, then after returning back to
the unmaximized frame, the frame is scaled down by 1 column and 1 line.

As a temporary workaround I bound F12 to the
resize-current-frame-to-default function which scales the window back to
its default dimensions.