Install whole-line-or-region package
Yet another gem from the Mickey Petersen's "Mastering Emacs."  This
package redefines C-w, M-w to operate on line when no region is active.
Thus, C-w without active region works exactly as C-S-<backspace>, but,
unlike C-S-<backspace>, works also in the terminal.  Some people do not
like this package, but it is worth to give it a chance.