Delete magit-section MELPA package
I have been using the magit-section package for a long time because I it
thought is somehow required, because each time I tried to remove it,
Magit stopped working properly.  Today, I realized that this was caused
by residual code in the native compilation cache.  The cache needs to be
pruned to after removal of the magit-section MELPA package and then it
works as expected.  In fact, the magit-section MELPA package is not
really needed in my setup, since all relevant Magit-related packages
are included in the emacs-magit RPM.