Set org-src-preserve-indentation as safe
When I use a code block for a source snippet that is supposed to be
copied to other file, I do not want the block to be indented.

The parameter is also practical if some syntax elements within the block
have to remain at the beginning of the line to be recognized.