Add variable to activate automatic date update
I have only a handful of org files where I want to update the date in
header.  I mostly want to keep it static.  To resolve this issue, I have
introduced a new variable:

  org-set-date-header-arg-as-time-stamp

This variable serves as a predicate to perform automatic update of the
date in the org-file header and it is nil by default.  User is expected
to set it up as a file variable for the specific files where is
the automatic update of the date upon saving desirable.