Fix automatic time-stamp update
It turned out that I need time stamp to be updated as part of the
before-save-hook only in specific files, where I can set it up using
file local variable.  This is the way how time stamps were intended to
be used according to the documentation, so it is better to follow the
design.  The following file local variable has to be set to activate
automatic time-stamp update for the file:

  eval: (add-hook 'before-save-hook 'time-stamp nil t)

I set time stamp format for org files, but since time-stamp call is not
globally added to the before-save-hook, by default the time-stamp is not
being updated before save.

One might always call the time-stamp function interactively to update
the time stamp to the current date if needed.