Install speed-type package
My keybr.com login link has expired and I see no point in paying for the
"lifetime access" again.  This might be a good opportunity to give a
chance to speed-type, which probably does not look as fancy, but it is
simple and efficient.  Furthermore, as I am getting more familiar with
Emacs, to make my standard workflow more efficient, it is recommendable
to slowly migrate more parts of the workflow into Emacs environment.