Set up zotxt mode
The zotxt turned up to be very useful integration and I want it to be
loaded automatically with org-mode.

I find the default bibliography style rather wordy.  Zotero comes with
"The Astronomy and Astrophysics Review" style which seems to be perfect
for what I need.  It took me a while to figure out that the correct
value of the zotxt-default-bibliography-style is the name of the XML
file without the extension (which is
"the-astronomy-and-astrophysics-review" in this case).